Move custom debugger logic out of the TypeScript adapter
