+++ packages/open-schema-type-script/src/example/custom-debugger/eventLogger.ts
