+++ packages/open-schema-type-script/src/example/exampleCustom.ts
@@ -1,7 +1,7 @@
 import { digikikify } from '../core/digikikify';
 import { blindCastEstinants } from './blindCastEstinants';
-import { eventLogger } from '../type-script-adapter/debugger/eventLogger';
-import { odeshinLogger } from '../type-script-adapter/debugger/odeshinLogger';
+import { eventLogger } from './custom-debugger/eventLogger';
+import { odeshinLogger } from './custom-debugger/odeshinLogger';
 import { fileAMentursection } from './file/fileAMentursection';
 import { fileAEstinant } from './file/fileA';
 import {
